requirements.txt
Django==1.11.15
Django==2.1
Django==3.2.5
djangorestframework==3.14.0
